requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appnope==0.1.3
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.2
bitarray==2.9.2
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.16.0
colbert-ai==0.2.15
comm==0.2.1
confection==0.1.4
configobj==5.0.8
configparser==6.0.0
cymem==2.0.8
damask @ file:///Users/mikerenwick/Documents/GitHub/damask
dataclasses-json==0.6.3
datasets==2.16.1
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
distro==1.9.0
etelemetry==0.3.1
executing==2.0.1
faiss-cpu==1.7.4
filelock==3.13.1
fitz==0.0.1.dev2
Flask==3.0.0
frozenlist==1.4.1
fsspec==2023.10.0
future==0.18.3
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
iniconfig==2.0.0
ipykernel==6.28.0
ipython==8.19.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.0
langchain-core==0.1.7
langcodes==3.3.0
langsmith==0.0.77
llama-index==0.9.26
looseversion==1.3.0
lxml==5.0.1
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nibabel==5.2.0
ninja==1.11.1.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.3
openai==1.6.1
packaging==23.2
pandas==2.1.4
parso==0.8.3
pathlib==1.0.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.3.0
preshed==3.0.9
prompt-toolkit==3.0.43
prov==2.0.0
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pyarrow-hotfix==0.6
pydantic==2.5.3
pydantic_core==2.14.6
pydot==2.0.0
Pygments==2.17.2
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyparsing==3.1.1
pystache==0.6.5
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyxnat==1.6
PyYAML==6.0.1
pyzmq==25.1.2
RAGatouille==0.0.2b0
rdflib==7.0.0
regex==2023.12.25
requests==2.31.0
ruff==0.1.11
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
sentence-transformers==2.2.2
sentencepiece==0.1.99
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
thinc==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
traits==6.3.2
transformers==4.36.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
urllib3==2.1.0
voyager==2.0.2
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
